move global definitions together and include them manually
