+++ lib/kleisli.rb
@@ -1,9 +1,3 @@
 require "kleisli/version"
-require "kleisli/maybe"
-require "kleisli/try"
-require "kleisli/future"
-require "kleisli/either"
-require "kleisli/composition"
-
-module Kleisli
-end
+require "kleisli/globals"
+include Kleisli::Globals
